Organize routes package
